# std.mem: add vectorized indexOfScalarPos and indexOfSentinel