@tiehuis 2023-09-15
While tested in the aforementioned repo we should add a few more tests in `std.mem` for these. Will add some page-crossing tests to confirm the less common branches are exercised.