@kcbanner 2023-09-17
Interesting, I will take a look at that MSVC failure