@kcbanner 2023-09-17
The CI failure wasn't actually a CBE issue, it was an issue compiling for the `x86_64-windows-msvc` target, which exercises a path where `indexOfSentinel` is called with T = `u16`. The panic was triggered inside the `windows_sdk` logic:

```
thread 10204 panic: incorrect alignment
C:\cygwin64\home\kcbanner\kit\zig\lib\std\mem.zig:1008:49: 0x7ff612ab4d03 in indexOfSentinel__anon_289748 (zig.exe.obj)
                    const block: *const Block = @ptrCast(@alignCast(p[i..][0..block_len]));
                                                ^
C:\cygwin64\home\kcbanner\kit\zig\lib\std\mem.zig:942:39: 0x7ff612ab40d3 in len__anon_289741 (zig.exe.obj)
                return indexOfSentinel(info.child, sentinel, value);
                                      ^
C:\cygwin64\home\kcbanner\kit\zig\lib\std\mem.zig:722:18: 0x7ff61254780c in span__anon_215251 (zig.exe.obj)
    const l = len(ptr);
                 ^
C:\cygwin64\home\kcbanner\kit\zig\src\windows_sdk.zig:281:46: 0x7ff611f298c0 in getString (zig.exe.obj)
            break :value_utf16le std.mem.span(value_utf16le_overestimated);
                                             ^
C:\cygwin64\home\kcbanner\kit\zig\src\windows_sdk.zig:129:61: 0x7ff61194977d in getString (zig.exe.obj)
        const value_utf16le = try registry_utf16le.getString(allocator, subkey_utf16le, value_name_utf16le);
                                                            ^
C:\cygwin64\home\kcbanner\kit\zig\src\windows_sdk.zig:357:67: 0x7ff6116b023d in find (zig.exe.obj)
            var path_maybe_with_trailing_slash = v10_key.getString(allocator, "", "InstallationFolder") catch |err| switch (err) {
                                                                  ^
C:\cygwin64\home\kcbanner\kit\zig\src\windows_sdk.zig:526:51: 0x7ff61142ffb1 in find (zig.exe.obj)
            const windows10sdk = Windows10Sdk.find(allocator) catch |err| switch (err) {
                                                  ^
C:\cygwin64\home\kcbanner\kit\zig\src\libc_installation.zig:200:56: 0x7ff6112437fd in findNative (zig.exe.obj)
            var sdk: ZigWindowsSDK = ZigWindowsSDK.find(args.allocator) catch |err| switch (err) {
                                                       ^
C:\cygwin64\home\kcbanner\kit\zig\src\Compilation.zig:4952:49: 0x7ff611243527 in detectLibCIncludeDirs (zig.exe.obj)
        libc.* = try LibCInstallation.findNative(.{ .allocator = arena, .verbose = true, .target = target });
                                                ^
C:\cygwin64\home\kcbanner\kit\zig\src\Compilation.zig:932:52: 0x7ff611289c8f in create (zig.exe.obj)
        const libc_dirs = try detectLibCIncludeDirs(
                                                   ^
C:\cygwin64\home\kcbanner\kit\zig\src\main.zig:3294:36: 0x7ff611381c58 in buildOutputType (zig.exe.obj)
    const comp = Compilation.create(gpa, .{
                                   ^
C:\cygwin64\home\kcbanner\kit\zig\src\main.zig:276:31: 0x7ff6111e1139 in mainArgs (zig.exe.obj)
        return buildOutputType(gpa, arena, args, .zig_test);
                              ^
C:\cygwin64\home\kcbanner\kit\zig\src\main.zig:214:20: 0x7ff6111de9de in main (zig.exe.obj)
    return mainArgs(gpa, arena, args);
                   ^
C:\cygwin64\home\kcbanner\kit\zig\lib\std\start.zig:473:101: 0x7ff6111de6bc in main (zig.exe.obj)
    return @call(.always_inline, callMainWithArgs, .{ @as(usize, @intCast(c_argc)), @as([*][*:0]u8, @ptrCast(c_argv)), envp });
                                                                                                    ^
???:?:?: 0x7ff613695625 in ??? (zig.exe)
???:?:?: 0x7ff61369568b in ??? (zig.exe)
???:?:?: 0x7ffb2a1d7613 in ??? (KERNEL32.DLL)
???:?:?: 0x7ffb2b2626b0 in ??? (ntdll.dll)
```

I think the mistake was using `block_len` as the alignment, when it should be `@alignOf(Block)`:

```diff
diff --git a/lib/std/mem.zig b/lib/std/mem.zig
index 9cb20fcbe..bd60e3e39 100644
--- a/lib/std/mem.zig
+++ b/lib/std/mem.zig
@@ -991,7 +991,7 @@ pub fn indexOfSentinel(comptime T: type, comptime sentinel: T, p: [*:sentinel]co
                         return i + std.simd.firstTrue(matches).?;
                     }

-                    i += std.mem.alignForward(usize, start_addr, block_len) - start_addr;
+                    i += (std.mem.alignForward(usize, start_addr, @alignOf(Block)) - start_addr) / @sizeOf(T);
                 } else {
                     // Would read over a page boundary. Per-byte at a time until aligned or found.
                     // 0.39% chance this branch is taken for 4K pages at 16b block length.
@@ -1003,7 +1003,7 @@ pub fn indexOfSentinel(comptime T: type, comptime sentinel: T, p: [*:sentinel]co
                     }
                 }

-                std.debug.assert(std.mem.isAligned(@intFromPtr(&p[i]), block_len));
+                std.debug.assert(std.mem.isAligned(@intFromPtr(&p[i]), @alignOf(Block)));
                 while (true) {
                     const block: *const Block = @ptrCast(@alignCast(p[i..][0..block_len]));
                     const matches = block.* == mask;
```