@andrewrk 2023-10-08
This is a bit unfortunate:

```
==1395995== Conditional jump or move depends on uninitialised value(s)
==1395995==    at 0x639CF39: mem.indexOfScalarPos__anon_7578 (mem.zig:1141)
==1395995==    by 0x63B820A: mem.indexOfScalar__anon_10770 (mem.zig:1100)
==1395995==    by 0x687A7DE: mem.lenSliceTo__anon_127951 (mem.zig:890)
==1395995==    by 0x65F553C: mem.sliceTo__anon_104438 (mem.zig:805)
==1395995==    by 0x63CFA39: os.getcwd (os.zig:2013)
==1395995==    by 0x63CF4A0: process.getCwdAlloc (process.zig:33)
==1395995==    by 0x65645FA: introspect.resolvePath (introspect.zig:123)
==1395995==    by 0x6568160: introspect.findZigLibDirFromSelfExe (introspect.zig:75)
==1395995==    by 0x6572615: main.cmdBuild (main.zig:4721)
==1395995==    by 0x6374F3F: main.mainArgs (main.zig:285)
==1395995==    by 0x6371C85: main.main (main.zig:213)
==1395995==    by 0x637169E: main (start.zig:486)
```

related: #17209

Not sure what to do about this yet. Also, this probably violates pointer provenance rules and is UB.